+++ Dependencies/Il2CppAssemblyGenerator/Config.cs
@@ -32,6 +32,7 @@ namespace MelonLoader.Il2CppAssemblyGenerator
             public string DeobfuscationRegex = null;
             public string UnityVersion = "0.0.0.0";
             public string DumperVersion = "0.0.0.0";
+            public string DumperSCRSVersion = "0.0.0.0";
             public bool UseInterop = true;
             public List<string> OldFiles = new List<string>();
         }